fix lỗi  UI còn lại của chương trình

+++ src/server_candidate/Source/app/models/candidate.m.js
@@ -295,11 +295,11 @@ module.exports = {
   
   addReviews: async (e, id_employer) => {
     const collection_reviews = await db.collection("reviews");
-    var rating=parseInt(e.star);
+    var rating=parseFloat(e.star);
     const collection_employer = await db.collection("employers").doc(id_employer).get();
     var rs=collection_employer.data();
     console.log(rs);
-    rating=(parseInt(rs.rating)*rs.list_reviews.length+rating)/(rs.list_reviews.length+1);
+    rating=(parseFloat(rs.rating)*rs.list_reviews.length+rating)/(rs.list_reviews.length+1);
     rating=rating.toFixed(1);
     console.log(rating);
     rating=parseFloat(rating);
